update with documentation

+++ src/pdata.h
@@ -5,7 +5,7 @@
 
 #include <Preferences.h>
 void pdata_init();
-void pdata_write(uint8_t what, uint8_t specific);
+void pdata_write(uint8_t what, uint16_t specific);
 void pdata_read();
 
 #endif
